# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STRESSNES luxury fashion ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all active collections
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"

    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Collection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get collection by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── Cart ────────────────────────────────────────────────────
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers (admin only)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Paginated customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCustomerList"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ── Reviews ─────────────────────────────────────────────────
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidationInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResult"

  # ── Inventory ────────────────────────────────────────────────
  /inventory/{variantId}:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get inventory for a product variant
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inventory status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryStatus"

    patch:
      operationId: updateInventory
      tags: [inventory]
      summary: Update inventory (admin only)
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUpdate"
      responses:
        "200":
          description: Inventory updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryStatus"

  # ── Newsletter ───────────────────────────────────────────────
...
components:
  schemas:
...
    Collection:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CollectionInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
        isActive:
          type: boolean
    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
        isActive:
          type: boolean

    # ── Cart ───────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, email, fullName, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PaginatedCustomerList:
      type: object
      required: [data, total, page, pageSize, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    CustomerUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        avatar:
          type: string

    # ── Reviews ────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, type, value, isActive]
      properties:
        id:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT]
        value:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
    CouponValidationInput:
      type: object
      required: [code, orderAmount]
      properties:
        code:
          type: string
        orderAmount:
          type: number
    CouponValidationResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        coupon:
          $ref: "#/components/schemas/Coupon"
        discountAmount:
          type: number
        message:
          type: string

    # ── Inventory ──────────────────────────────────────────────
    InventoryStatus:
      type: object
      required: [variantId, stockQty, lowStockThreshold]
      properties:
        variantId:
          type: string
        stockQty:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        lastUpdated:
          type: string
          format: date-time
    InventoryUpdate:
      type: object
      required: [change, reason]
      properties:
        change:
          type: integer
        reason:
          type: string
          enum: [PURCHASE, RETURN, ADJUSTMENT, RESERVATION, RELEASE, DAMAGE, RESTOCK]
        note:
          type: string

    # ── Newsletter ─────────────────────────────────────────────
...